+++ packages/jest-environment-node/src/index.ts
@@ -9,7 +9,10 @@ import {Context, Script, createContext, runInContext} from 'vm';
 import {Config, Global} from '@jest/types';
 import {ModuleMocker} from 'jest-mock';
 import {installCommonGlobals} from 'jest-util';
-import {JestFakeTimers as FakeTimers} from '@jest/fake-timers';
+import {
+  JestFakeTimers as LegacyFakeTimers,
+  LolexFakeTimers,
+} from '@jest/fake-timers';
 import {JestEnvironment} from '@jest/environment';
 
 type Timer = {
@@ -20,7 +23,8 @@ type Timer = {
 
 class NodeEnvironment implements JestEnvironment {
   context: Context | null;
-  fakeTimers: FakeTimers<Timer> | null;
+  fakeTimers: LegacyFakeTimers<Timer> | null;
+  fakeTimersLolex: LolexFakeTimers | null;
   global: Global.Global;
   moduleMocker: ModuleMocker | null;
 
@@ -70,25 +74,28 @@ class NodeEnvironment implements JestEnvironment {
       refToId: timerRefToId,
     };
 
-    this.fakeTimers = new FakeTimers({
+    this.fakeTimers = new LegacyFakeTimers({
       config,
       global,
       moduleMocker: this.moduleMocker,
       timerConfig,
     });
+
+    this.fakeTimersLolex = new LolexFakeTimers({config, global});
   }
 
-  setup() {
-    return Promise.resolve();
-  }
+  async setup() {}
 
-  teardown() {
+  async teardown() {
     if (this.fakeTimers) {
       this.fakeTimers.dispose();
     }
+    if (this.fakeTimersLolex) {
+      this.fakeTimersLolex.dispose();
+    }
     this.context = null;
     this.fakeTimers = null;
-    return Promise.resolve();
+    this.fakeTimersLolex = null;
   }
 
   // TS infers the return type to be `any`, since that's what `runInContext`